Encapsulate Timers and Analog measurements
This commit encapsulates timers and ADC in their specific locations.
Now the main only initializes the Logger which initialize the channels.
