+++ usr/user.mk
@@ -6,3 +6,5 @@ USERINC =  $(USERLIB) \
            
 # List of all the Userlib files
 USERSRC =  $(USERLIB)/src/logger.c
+USERSRC +=  $(USERLIB)/src/logger_analog_ch.c
+USERSRC +=  $(USERLIB)/src/logger_timing.c
